Convert from timer- to event-based updates
Previously a timer was used to refresh the contents of the line, along with a
hook to watch for window updates.  This might be a poor approach, since the line
is continually updated, even if there is no reason to update.

With the debounced update function we can rely solely on hooks to update the
line.  Even when attached to the `post-command-hook', the performance hit should
be negligible, and even considerably less than that of the timer based approach,
since in the situation of update attempts in quick succession, only the timer
interval is recalculated.